Moved and fixed broken links of master_table
