+++ addie/processing/mantid/master_table/import_from_database/conflicts_solver.py
