+++ addie/processing/mantid/master_table/import_from_database/format_json_from_database_to_master_table.py
@@ -3,7 +3,7 @@ from __future__ import (absolute_import, division, print_function)
 import collections
 import copy
 
-from addie.master_table.master_table_loader import FormatAsciiList
+from addie.processing.mantid.master_table.master_table_loader import FormatAsciiList
 from addie.utilities.list_runs_parser import ListRunsParser
 from addie.utilities.general import json_extractor
