+++ addie/processing/mantid/master_table/import_from_database/global_rule_handler.py
