+++ addie/processing/mantid/master_table/master_table_loader.py
@@ -12,8 +12,8 @@ from qtpy import QtCore, QtGui
 from addie.utilities.file_handler import FileHandler
 from addie.utilities.list_runs_parser import ListRunsParser
 from addie.utilities.set import Set
-from addie.master_table.table_row_handler import TableRowHandler
-from addie.master_table.utilities import LoadGroupingFile
+from addie.processing.mantid.master_table.table_row_handler import TableRowHandler
+from addie.processing.mantid.master_table.utilities import LoadGroupingFile
 
 from addie.icons import icons_rc
