+++ addie/processing/mantid/master_table/periodic_table/chemical_formula_handler.py
