+++ addie/processing/mantid/master_table/reduction_configuration_handler.py
@@ -5,7 +5,7 @@ from addie.utilities import load_ui
 from qtpy import QtGui, QtCore
 
 from addie.processing.mantid.make_calibration_handler.make_calibration import MakeCalibrationLauncher
-from addie.master_table.utilities import LoadGroupingFile
+from addie.processing.mantid.master_table.utilities import LoadGroupingFile
 
 from addie.icons import icons_rc
