+++ addie/processing/mantid/master_table/utilities.py
